+++ apps/build-onchain-apps/onchainKit/hooks/useBlockExplorerLink.test.ts
@@ -5,7 +5,7 @@
 import { useNetwork } from 'wagmi';
 import { Chain } from 'viem/chains';
 import { renderHook } from '@testing-library/react';
-import useBlockExplorerLink, { HashType } from './useBlockExplorerLink';
+import { HashType, useBlockExplorerLink } from './useBlockExplorerLink';
 
 jest.mock('wagmi', () => ({
   useNetwork: jest.fn(),
