+++ apps/build-onchain-apps/onchainKit/index.ts
@@ -1,1 +1,3 @@
-export { AccountConnectButton } from './AccountConnectButton';
+export { OnchainAvatar } from './components/OnchainAvatar';
+export { useCollectionMetadata } from './hooks/useCollectionMetadata';
+export { useBlockExplorerLink } from './hooks/useBlockExplorerLink';
