+++ apps/build-onchain-apps/package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "@coinbase/onchainkit": "^0.1.0",
     "@coinbase/wallet-sdk": "^3.9.1",
+    "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-navigation-menu": "^1.1.4",
     "@rainbow-me/rainbowkit": "1.3.3",
